requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.4.0; python_version >= '3.8'
ariadne==0.23.0
bcrypt==4.1.3
certifi==2024.2.2; python_version >= '3.6'
cffi==1.16.0; platform_python_implementation != 'PyPy'
click==8.1.7; python_version >= '3.7'
cryptography==42.0.7
dnspython==2.6.1; python_version >= '3.8'
email-validator==2.1.1; python_version >= '3.8'
fastapi==0.111.0; python_version >= '3.8'
fastapi-cli==0.0.4; python_version >= '3.8'
fastapi-limiter==0.1.6; python_version >= '3.9' and python_version < '4.0'
fastapi-route-logger-middleware==0.2.3; python_version >= '3.7'
graphql-core==3.2.3; python_version >= '3.6' and python_version < '4'
greenlet==3.0.3; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.5; python_version >= '3.8'
httptools==0.6.1
httpx==0.27.0; python_version >= '3.8'
idna==3.7; python_version >= '3.5'
jinja2==3.1.4; python_version >= '3.7'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
mdurl==0.1.2; python_version >= '3.7'
orjson==3.10.3; python_version >= '3.8'
passlib[bcrypt]==1.7.4
psycopg[binary]==3.1.19; python_version >= '3.7'
psycopg-binary==3.1.19
pycparser==2.22; python_version >= '3.8'
pydantic==2.7.1; python_version >= '3.8'
pydantic-core==2.18.2; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
pyjwt[crypto]==2.8.0; python_version >= '3.7'
python-decouple==3.8
python-dotenv==1.0.1; python_version >= '3.8'
python-multipart==0.0.9; python_version >= '3.8'
pyyaml==6.0.1
redis==5.1.0b5; python_version >= '3.8'
rich==13.7.1; python_full_version >= '3.7.0'
shellingham==1.5.4; python_version >= '3.7'
sniffio==1.3.1; python_version >= '3.7'
sqladmin==0.17.0; python_version >= '3.8'
sqlalchemy==2.0.30; python_version >= '3.7'
starlette==0.37.2; python_version >= '3.8'
typer==0.12.3; python_version >= '3.7'
typing-extensions==4.12.0; python_version >= '3.8'
ujson==5.10.0; python_version >= '3.8'
uvicorn[standard]==0.30.0; python_version >= '3.8'
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
wtforms==3.1.2; python_version >= '3.8'
